cambios

+++ src/main/java/com/proyectofinal/inventario/repository/RepositorioProveedor.java
@@ -3,5 +3,5 @@ package com.proyectofinal.inventario.repository;
 import com.proyectofinal.inventario.entity.Proveedor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RepositorioProveedor extends JpaRepository<Proveedor,Integer> {
+public interface RepositorioProveedor extends JpaRepository<Proveedor,Long> {
 }
